requirements.txt
docx2txt==0.8
html5lib==1.1
beautifulsoup4==4.12.3
nbformat==5.10.4
openpyxl==3.1.5
PyMuPDF==1.24.4
python-docx==1.1.2
python-pptx==1.0.2
fpdf==1.7.2
markdown==3.7
pylatexenc==2.10
llama-index==0.11.11
llama-index-embeddings-huggingface==0.3.1
llama-index-llms-ollama==0.3.2
python-dotenv==1.0.1
# for streamlit
streamlit==1.37.1
ollama